first stage p2p v1.1

+++ services/tickets.service.js
@@ -5,9 +5,10 @@ export async function listTickets(userid) {
     return tickets
 }
 
-export async function cancelTicket(ticket_id) {
+export async function cancelTicket(ticket_id, userid) {
     const ticket = await (await conexion.query('SELECT * FROM tickets WHERE ticket_id=($1)', [ticket_id])).rows[0]
     if (ticket.status !== 'pending') return { status: false, content: 'Ticket already matched' }
+    if (ticket.owner !== userid) return { status: false, content: "You are not the owner of this ticket" }
     await conexion.query('DELETE FROM tickets WHERE ticket_id=($1)', [ticket_id])
     return { status: true, content: 'Ticket successfully cancelled' }
 }
